NF: implement a skeleton universe incl. layout
Tried to write the universe, but noticed that I need to be able to specify the
universe-layout somehow. I basically used a similar method as before. The layout
is specified as a string with special symbols. I also do some extensive legality
checking here to avoid future fuckups.

Also the simple executions should be refactored into unit-tests.