RF: de-globalize legal layout chars
The layout_chars were previoulsy global. Now they are stored in the CTFUniverse.
The idea is that each universe can specify its own layout chars. This had some
serious reprecussions on the Layout class, since it needed to be parametrized.

Also the extract_food and initial_poistions were moved from module level to
CTFUniverse class for compactness. I made them staticmethods so that the testing
would be easier. Can of course be made instance methods too, if you like.