RF: Major refactor of Datamodel
Instead of keeping track of bots, their zones, their initial_positions, we now
have a Bot and a Team class.

While this first approach looks not to bad, there are some issues here.

The main problem is that Team and Bot are tightly coupled by mutual composition.
I.e. Team has a ref to all Bots, and the Bot has a ref to the team it is on. The
first major issue i ran into, was that __eq__() failed due to exceeding maximum
recursion depth. This is obvious, since  naive implementation of __eq__() simply
compares the __dict__ of self and other. Thereby the __eq__() of Team called the
__eq__() of Bot calles .... And so on. One way to alleviate this is to modify
the __eq__() of Bot to compare only some aspects of the Team e.g. the name.
We could do the same thing for __repr__() obviously, but this causes headaches,
because the Bot relies on information in the Team to make decisions about which
state to initialise itself in. If that information is missing, Exceptions are
raised, and thus the 'eval(bot.__repr__())' is completely broken.

So, after this iterative exploration, the verdict on mutual composition is more
of a nay.

We need to rethink about having a Team class at all. Maybe its better to have a
dict of team_name -> bot_list, and store the homezone of the bots inside the
bots themselves. Scorekeeper would be again the universe.

An alternative we could go for is to have only a Team class which holds
information about the Bots, lits of current and initial positions, states etc..
-- but this would be no better than storing this information directly in the
universe itself -- bringing us back to square one.