RF: Reverse Mesh.height and Mesh.width.
Previously, Mesh was considered a matrix with mesh[i,j] referencing
the i-th column and the j-th row. This meant, that the first index was
the height coordinate while the second specified the width.

However, in our case a Mesh is rather a mapping from a two-dimensional
coordinate system NxN into object space and positions on the Mesh
should be specified with a (x, y) tuple and the shape should be given
by (width, height).

Internally, the x coordinate is the one which runs faster, so (0, 1)
comes *after* (1, 0). This makes it slightly easier to define layouts.
The order, however, should not be taken for granted when using the
Mesh. All access should be done using the tuple notation or through
the provided iterators.