BF/UT/ENH: fix some subtle bugs in demo players
The BFSPlayer was crashing after having been eaten, because it did not check for
being reset, and then tried to make illegal moves. We now check for destruction
and then reset the path.

The NQRandom player was crashing because it had moved into a dead-end and then
removed the previous position from its list of possible moves, thereby
random.choice raised an exception due to an empty list. We now remove stop
first, and then remove the previous_position, only if there are other
alternatives.