BF/ENH: It was (basically) working ALL THE TIME! I curse you, you stupid, curious delegation pattern.
What was happening? While animations did visibly update and draw their
circles on the canvas, they would not vanish after the next animation
step had been drawn. Calling update_idletasks a myriad of times to
trigger updating the canvas would not help. Neither did forceful
deletion of the items help. Why did it not help? Because drawing of the
items was done using the object id() as a tag. This meant that the id
could be retrieved from the object easily and deletion would remove one
object only from the canvas. The curious delegation pattern had it that
it delegated its drawing tasks to another instance, which was
instantiated with a copy of the original class’s attributes. Everything
worked fine only that all newly created canvas items were created using
the id() of the delegation class. As this id() was being thrown away
after each step, there was no way that the canvas items could ever be
found again. And they lived happily ever after. (Or at least until the
next delete(ALL) would become their destiny.)