BF/ENH/UT: fix BFSPlayer
There was a bug in the bfs, such that if no food was found a bogus path was
returned. We fix this bug and instead return a path of length one consisting of
the current position only. Thus if the Player cannot find food it simply stands
still. This is still a bit dirty, but will refactor the algorithm when moving
this code to the graph module.