NF: Removing the actor based server setup in favour of a zmq based connection.
This change is two-fold. For one it greatly straightens the way
our connections work and speeds up the context switching through
the removal of slow Python 2.7 thread queues. On the other side
it also gets rid of the local–non-local distinction by sending
all messages through zmq sockets.

It is assumed that every client connects to the server through
a zmq.PAIR connection (the address of which is pre-defined). It
is also assumed that per connection the server may only send
one request at a time. Multiple requests are silently discarded
(as opposed to the old model which was more robust in this regard).