ENH: relax checks in noiser tests
Tests checked if every position is hit at least 5 times. This caused
at least one test to fail in every second or third run on my laptop.
Just check if every position is hit at least 1 time.

To simplify the tests, explicit positions are not checked, just that
there is the expected number of positions hit.