UT: Check that bind_socket fails on a bad socket.
What we really want to accomplish with this check, though, is finding
out about API changes of pyzmq. Starting with pyzmq 13, the long forms
of `zmq.core.error.ZMQError` (and similar) do not seem to work anymore.
A solution is to use the much simpler `zmq.ZMQError` which is also in
pyzmw 2.2.0 and which even might have been there forever. Who knows. The
next commit will then fix all of the longer forms to their correct short
form. (We’re even cheating in this commit by using the short form in the
test case.)