RF: Refactoring pelitagame. (wip)
First step in cleaning up ‘pelitagame’ while adding more and more features.

This commit also adds a `--replay` option which makes the standalone ‘pelitaviewer’ obsolete. As a plus, we’re able to directly use the ‘pelitagame’ arguments for setting the fps. (Future work might add a controller so we can step through events.)