ENH: Preliminary support for module runner specification
This commit implements a prefix syntax in the module specification which
allows pelitagame to automatically start a player with a specific
executable. Currently, only Python executables are hard-coded but at
some point this may be made more flexible.

Example:

  $ python3 pelitagame py2@old_school_player py3@modern_player

This starts pelitagame in Python 3 mode which in turn runs subprocesses
of Python 2 and Python 3 for the individual modules.

If no prefix is given, this defaults to py@, which specifies the same
Python interpreter that was used to run pelitagame itself.