ENH: Rework the import logic in pelita_player
Previously, our check_module function would take a guess at what a
proper module would look like and fail if it looked differently.  We now
relax some of the constraints and just try to import the module (with
importlib). If it works, good, if it doesn’t, then we can still fail.

As a side-effect, this change allows to specify a module without path
delimiters (unless it would collide with a built-in player) and modules
containing only .pyc files.