BF: Fix broken replay mode.
The ReplayPublisher now includes controller functionality, which means
that the Tk viewer can now request the next frame, when it wants to.
This means that we can get rid of the more or less arbitrary sleep
functions for proper synchronisation. Additionally, we can now pause the
replay and step through it. (Stepping through rounds is not implemented
yet.)