RF: Store universe in GameMaster.game_state and remove it from APIs
We store the dict representation of CTFUniverse in the game_state. Since
the universe is only accessed in a few methods in GameMaster, we can
live with a small number of conversions. The worst drawback, however, is
that we use GameMaster.universe extensively in the tests, where
conversions are more annoying but hopefully only temporary.

Removing the universe arguments in the AbstractViewer and AbstractTeam
APIs should make integration with the new core rewrite easier.
The API for SimpleTeam or AbstractPlayer have not been changed yet.