ENH: Add dataclass for the GameState
This replaces the informative canonical_keys with a more normative
dictionary layout. Currently, the GameState class is only used in
setup_game to create the game_state dict, which is then passed around
instead. Test functions, however, can use the dataclass to check that
all other relevant functions of game are passing around complete (at
least considering the keys) game state dicts.

This class is only an experiment for now and may or may not be extended
with more helper functions or be removed completely in the future.