RF: Revised the noiser API
The plan is to do the noising inside the `prepare_bot_state` function,
which takes a complete game_state and shapes it into a form that can be
sent to a player. This means, we don’t have to rely on the whole
game_state here but can simply use the current bot’s position and a list
of enemy bots that can potentially be noised.