RF: Change game_state['error'] to list of dicts.
There is only one error per (round, turn) allowed and all lead to the
same outcome -> a random move. Mirroring this in the data structure
simplifies the checks.

Currently, the turn is in range(0, 3), ie. global, even though
the error state is split between the two teams and therefore we could
also use the team-internal turn.

Additionally, this fixes the double-error problem when a timeout occurred
and a user would get an additional error for making a wrong move.