BF: Make msg_id handling in zmq connections explicit
The uuid is used to match a reply to its request. Storing it inside the
ZMQConnection instance turned out to be error-prone as it could be
overwritten by a belated answer to an older reply.

We now require a caller to include the uuid in the recv method so that
the correct answer is returned.