BF: Remote players will write to the folder given in --store-output
Note that the new Python process needs to run in PYTHONUNBUFFERED mode,
as Python’s print function will not automatically flush when not writing
to stdout/stderr. Hence, we need to force unbuffered writing, as we are
mostly interested in the very last messages that a process writes (ie.
an error message).