use list of tuples for game_state list of positions consistently in the TK UI.
By doing this conversion we are fixing several little bugs in the status line
showed when clicking on a field. With this we remove from the user facing
interface the last leaks of the internal bot index. Now bots are identified by
their color and their index within the respective team.