NF: add flag to raise Exception in run_game...
When using run_game in unit tests or in background games, one typically
wants Exceptions generated by the move functions to be immediately raised
and not captured by the game master. Otherwise unit tests may look like
they pass, while they are indeed failing. Without this one would have to
check the contents of the errors and fatal_errors dict in game_master,
but that loses a bit of details and may be too late...

Open for comments and/or improvements