Promote pytest and networkx to depenencies for pelita
pytest is needed for the tests also in pelita_template, so it makes
sense if it gets installed unconditionally together with pelita.

networkx is not strictly a dependency, but it makes sense to use it
in pelita_template, and it is indeed used already in some demos, so
let's have it also for pelita, given that pelita_template will come
back to the main repo at some point.