BF: fix extremely long-standing bug in maze_generator!
Since the conversion to Python3 and networkx away from Python 2 and
mdp.Graph done with commit 7603cdbc9f96a7e3dbf1 , the dead_ends removal
functionality was completely broken. Basically the function
nx.bfs_successors(graph, start_node) was *not* exploring all nodes in
the graph, and thus was feeding a shorter list of nodes to the functions
that were to detect dead ends. By looping through the whole graph
instead of choosing a particular order we can reproduce exactly the
mazes produced with the code before commit 7603cdbc9f96a7e3dbf1 . These
mazes are much nicer!

On a side note, this code is still using a directed graph, because it
tries to replicate the Graph usage from mdp.Graph, which was a directed
graph implementation. Next step will be to remove this and just use
a more simple undirected Graph.

This was quite a journey among 8 years old Pythons and ancient Pelitas
and networkxes.