removed unused create_all_mazes.sh script and move detect_dead_ends.py
An example of how to re-create all built-in pelita layouts has been
added to the help of the installed script pelita-createlayout.

The check_layout_consistency.py script is more of an emergency script
that can be used when there are doubts about some layout. I used it
a lot during the development of the new maze generation framework, but
I don't think it is useful enough to be installed by default. It can
stay in the repo on the top-level directory.