Do not return state from move()
Things are simpler if we always initialize the state to be an empty
dictionary, and simply pass the same mutable object always to the move
function. In cases where the player does not actually use the state to
store anything, it can ignore the variable completely. The object also
has always the same type (dict), so there is less chance of trying to
a TypeError.