ENH: Use runpy for importing a player.
Previously, we used our own hack for importing a user’s Python file for a
player, which involved modifying sys.path to include the directory of
the Python file and then import the Python file as a module with
importlib.import_module. This came with some mild (and sometimes
unexpected) annoyances, namely, it was not possible to import a file
from a module with the same name of a previously imported module.

A player in a directory called ‘sys’, or – more likely – ‘pelita’ would
therefore never be found by the program. It was also not possible to
import two players from different but equally-named folders, although
this problem had been circumvented by running each player in a single
subprocess always.

This commit changes the import mechanism to using the runpy module for
importing the player’s module, TEAM_NAME and move function.