Merge pull request #708 from Debilski/feature/dimensions-in-layout
We added a "shape" attribute to the layout and state dicts, so that we do not
need to run the costly (and error-prone) max(walls) + 1 trick for estimating
height and width of the maze.

Additionally, the wall coordinates are now stored in a set in order to
optimise the check whether a position is valid. 6cdc1f08674a1a998eff72297531b6e1026b67a5