this is a failing test that reproduces an issue we...
... observed in Mexico City:

Running the code wiht the pelita CLI shows the right error:

Could not load test_broken.py: invalid syntax (doesnotexistbelieveme.py, line 1)): SyntaxError
...
[and then a nice traceback with the pointer to the file and the line
with the SyntaxError]

In the test we get instead a different thing:

AttributeError: module 'teamy' has no attribute 'move'

--> we have observed this "wrong" error reporting also with the CLI, but
we can't reproduce it right now. My guess is that we have to go through
the whole error-catching on import logic and fix it/streamline it.