RF: Handle multiple pelita players in a single pelita server.
We change the old remote:tcp://ip:port schema for connecting to a pelita
server to use our own url scheme: pelita://ip/path (using a default
port). The supplied path is used to select a team on the server.

We switch to click for the server command line.

TODO: zeroconf for multiple players and API-based management controls