BF: Fix server getting slower with every new match
ZMQ (or the OS) gets a little slower with every additional socket to
watch. It isn’t a lot but it adds up for ~300x2x2 messages each game
that go through our poll.

Interestingly, it is already noticable with just a few additional
sockets, which would happen during a regular server run. Any decent
high-perfomance implementation of the pelita server should perhaps poll
the ROUTER sockets separately from the PAIR sockets (especially, if we
are just forwarding what we receive on the PAIR sockets) or replace the
PAIR–PAIR messaging in the backend players with a similar ROUTER–DEALER
model.