RF: Send final game state with to network players
This makes it easier for backends/middleends to capture the last state
that a player receives without having to parse it and ensure that it
actually contains a state.

A corollary to not doing that kind of deep package inspection is that,
once a game has been started, we cannot distinguish between control
messages and game info (set_initial/get_move) messages anymore and have
to assume that everything is a message that needs to be forwarded to the
backend player (unless we change our ZMQ protocoll to include an
additional control channel).