RF: Move the ‘bot welcome message’ to the main game
Previously, a pelita_player process would print its name when it was
started. Apart from giving a normal user the name info, it also served
as a tiny debugging hint in that it would tell a Pelita developer that
the player process had in fact started (and, more subtly, how long the
startup took).

However, this also required telling pelita_player on the command line
about its colour and necessitated an additional flag that told the
player to be quiet (when used on a server for example where this info
was not useful). In addition, games against a remote player would then
not show this info to a normal user.

This commit moves the welcome messages to the main game. Further work
needs to be done to properly give debugging hints for
slow/non-functional players.